Convert ChainDB to a trait implemented by Hammersbald struct. Make hammersbald default optional dependency.
