+++ Cargo.toml
@@ -11,6 +11,9 @@ keywords = [ "bitcoin" ]
 readme = "README.md"
 edition = "2018"
 
+[features]
+default = ["hammersbald"]
+
 [lib]
 name = "murmel"
 path = "src/lib.rs"
@@ -19,7 +22,6 @@ path = "src/lib.rs"
 lightning = { version ="0.0.9", optional=true }
 bitcoin = { version= "0.21", features=["use-serde"]}
 bitcoin_hashes = "0.7"
-hammersbald = { version= "2.4", features=["bitcoin_support"]}
 mio = "0.6"
 rand = "0.7"
 log = "0.4"
@@ -30,6 +32,9 @@ futures-preview = "=0.3.0-alpha.18"
 futures-timer = "0.3"
 serde="1"
 serde_derive="1"
+
+## optional
+hammersbald = { version= "2.4", features=["bitcoin_support"], optional=true }
 
 [dev-dependencies]
 rustc-serialize = "0.3"
